+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/experimental/plugin/listeners/HideChannelListener.kt
@@ -1,0 +1,56 @@
+package io.getstream.chat.android.client.experimental.plugin.listeners
+
+import io.getstream.chat.android.client.ChatClient
+import io.getstream.chat.android.client.utils.Result
+import io.getstream.chat.android.core.ExperimentalStreamChatApi
+
+@ExperimentalStreamChatApi
+/**
+ * Listener of [ChatClient.hideChannel] requests.
+ */
+public interface HideChannelListener {
+
+    /**
+     * Run precondition for the request. If it returns [Result.isSuccess] then the request is run otherwise it returns
+     * [Result.error] and no request is made.
+     *
+     * @param channelType Type of the requested channel.
+     * @param channelId Id of the requested channel.
+     * @param clearHistory Boolean, if you want to clear the history of this channel or not.
+     *
+     * @return [Result.success] if precondition passes otherwise [Result.error]
+     */
+    public suspend fun onHideChannelPrecondition(
+        channelType: String,
+        channelId: String,
+        clearHistory: Boolean
+    ): Result<Unit>
+
+    /**
+     * Runs side effect before the request is launched.
+     *
+     * @param channelType Type of the requested channel.
+     * @param channelId Id of the requested channel.
+     * @param clearHistory Boolean, if you want to clear the history of this channel or not.
+     */
+    public suspend fun onHideChannelRequest(
+        channelType: String,
+        channelId: String,
+        clearHistory: Boolean
+    )
+
+    /**
+     * Runs this function on the result of the request.
+     *
+     * @param result Result of this request.
+     * @param channelType Type of the requested channel.
+     * @param channelId Id of the requested channel.
+     * @param clearHistory Boolean, if you want to clear the history of this channel or not.
+     */
+    public suspend fun onHideChannelResult(
+        result: Result<Unit>,
+        channelType: String,
+        channelId: String,
+        clearHistory: Boolean,
+    )
+}